+++ abci/code/code.go
@@ -104,5 +104,7 @@ const (
 	ProxyNodeIsNotActive                      uint32 = 77
 	NodeIDInIdPListIsNotActive                uint32 = 78
 	NodeIDInASListIsNotActive                 uint32 = 79
+	RoleIsNotAS                               uint32 = 80
+	RequestIsNotClosed                        uint32 = 81
 	UnknownError                              uint32 = 999
 )