+++ abci/did/v1/dataStruct.go
@@ -719,3 +719,16 @@ type GetAccessorsInAccessorGroupParam struct {
 type GetAccessorsInAccessorGroupResult struct {
 	AccessorList []string `json:"accessor_list"`
 }
+
+type RevokeAccessorMethodParam struct {
+	AccessorIDList []string `json:"accessor_id_list"`
+	RequestID      string   `json:"request_id"`
+}
+
+type GetAccessorOwnerParam struct {
+	AccessorID string `json:"accessor_id"`
+}
+
+type GetAccessorOwnerResult struct {
+	NodeID string `json:"node_id"`
+}